changed rec_num to ulong

+++ src/task/assoc/createartasassociationsjob.h
@@ -32,7 +32,7 @@ struct UniqueARTASTrack
 {
     int utn;
     int track_num;
-    std::map<int, std::pair<std::string, boost::posix_time::ptime>> rec_nums_tris_;  // rec_num -> (tri, timestamp)
+    std::map<unsigned long, std::pair<std::string, boost::posix_time::ptime>> rec_nums_tris_;  // rec_num -> (tri, timestamp)
     // std::vector<int> track_nums_;
     boost::posix_time::ptime first_ts_;
     boost::posix_time::ptime last_ts_;
@@ -86,11 +86,12 @@ protected:
     std::map<int, UniqueARTASTrack> finished_tracks_;  // utn -> unique track
 
     // dbo -> hash -> rec_num, timestamp
-    std::map<std::string, std::multimap<std::string, std::pair<int, boost::posix_time::ptime>>> sensor_hashes_;
+    std::map<std::string,
+        std::multimap<std::string, std::pair<unsigned long, boost::posix_time::ptime>>> sensor_hashes_;
 
     std::map<std::string,
-        std::map<unsigned int,
-            std::tuple<unsigned int, std::vector<std::pair<std::string, unsigned int>>>>> associations_;
+        std::map<unsigned long,
+            std::tuple<unsigned int, std::vector<std::pair<std::string, unsigned long>>>>> associations_;
     // dbcontent -> rec_num -> <utn, src rec_nums (dbcontent, rec_num)>
